+++ test/Npgsql.Tests/Types/JsonDynamicTests.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Text.Json;
-using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Npgsql.Properties;
@@ -15,64 +14,6 @@ namespace Npgsql.Tests.Types;
 [TestFixture(MultiplexingMode.Multiplexing, NpgsqlDbType.Jsonb)]
 public class JsonDynamicTests : MultiplexingTestBase
 {
-    [Test]
-    public Task Roundtrip_JsonObject()
-        => AssertType(
-            new JsonObject { ["Bar"] = 8 },
-            IsJsonb ? """{"Bar": 8}""" : """{"Bar":8}""",
-            PostgresType,
-            NpgsqlDbType,
-            // By default we map JsonObject to jsonb
-            isDefaultForWriting: IsJsonb,
-            isDefaultForReading: false,
-            isNpgsqlDbTypeInferredFromClrType: false,
-            comparer: (x, y) => x.ToString() == y.ToString());
-
-    [Test]
-    public Task Roundtrip_JsonArray()
-        => AssertType(
-            new JsonArray { 1, 2, 3 },
-            IsJsonb ? "[1, 2, 3]" : "[1,2,3]",
-            PostgresType,
-            NpgsqlDbType,
-            // By default we map JsonArray to jsonb
-            isDefaultForWriting: IsJsonb,
-            isDefaultForReading: false,
-            isNpgsqlDbTypeInferredFromClrType: false,
-            comparer: (x, y) => x.ToString() == y.ToString());
-
-    [Test]
-    [IssueLink("https://github.com/npgsql/npgsql/issues/4537")]
-    public async Task Write_jsonobject_array_without_npgsqldbtype()
-    {
-        // By default we map JsonObject to jsonb
-        if (!IsJsonb)
-            return;
-
-        await using var conn = await OpenConnectionAsync();
-        var tableName = await TestUtil.CreateTempTable(conn, "key SERIAL PRIMARY KEY, ingredients json[]");
-
-        await using var cmd = new NpgsqlCommand { Connection = conn };
-
-        var jsonObject1 = new JsonObject
-        {
-            { "name", "value1" },
-            { "amount", 1 },
-            { "unit", "ml" }
-        };
-
-        var jsonObject2 = new JsonObject
-        {
-            { "name", "value2" },
-            { "amount", 2 },
-            { "unit", "g" }
-        };
-
-        cmd.CommandText = $"INSERT INTO {tableName} (ingredients) VALUES (@p)";
-        cmd.Parameters.Add(new("p", new[] { jsonObject1, jsonObject2 }));
-        await cmd.ExecuteNonQueryAsync();
-    }
-
     [Test]
     public async Task As_poco()
         => await AssertType(
