+++ test/Npgsql.Tests/Types/JsonTests.cs
@@ -1,10 +1,8 @@
 using System;
-using System.Data;
 using System.IO;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
-using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using NpgsqlTypes;
 using NUnit.Framework;
@@ -166,6 +164,64 @@ public class JsonTests : MultiplexingTestBase
         }
 
         Assert.That(car.RootElement.GetProperty("key").GetString(), Is.EqualTo("foo"));
+    }
+
+    [Test]
+    public Task Roundtrip_JsonObject()
+        => AssertType(
+            new JsonObject { ["Bar"] = 8 },
+            IsJsonb ? """{"Bar": 8}""" : """{"Bar":8}""",
+            PostgresType,
+            NpgsqlDbType,
+            // By default we map JsonObject to jsonb
+            isDefaultForWriting: IsJsonb,
+            isDefaultForReading: false,
+            isNpgsqlDbTypeInferredFromClrType: false,
+            comparer: (x, y) => x.ToString() == y.ToString());
+
+    [Test]
+    public Task Roundtrip_JsonArray()
+        => AssertType(
+            new JsonArray { 1, 2, 3 },
+            IsJsonb ? "[1, 2, 3]" : "[1,2,3]",
+            PostgresType,
+            NpgsqlDbType,
+            // By default we map JsonArray to jsonb
+            isDefaultForWriting: IsJsonb,
+            isDefaultForReading: false,
+            isNpgsqlDbTypeInferredFromClrType: false,
+            comparer: (x, y) => x.ToString() == y.ToString());
+
+    [Test]
+    [IssueLink("https://github.com/npgsql/npgsql/issues/4537")]
+    public async Task Write_jsonobject_array_without_npgsqldbtype()
+    {
+        // By default we map JsonObject to jsonb
+        if (!IsJsonb)
+            return;
+
+        await using var conn = await OpenConnectionAsync();
+        var tableName = await TestUtil.CreateTempTable(conn, "key SERIAL PRIMARY KEY, ingredients json[]");
+
+        await using var cmd = new NpgsqlCommand { Connection = conn };
+
+        var jsonObject1 = new JsonObject
+        {
+            { "name", "value1" },
+            { "amount", 1 },
+            { "unit", "ml" }
+        };
+
+        var jsonObject2 = new JsonObject
+        {
+            { "name", "value2" },
+            { "amount", 2 },
+            { "unit", "g" }
+        };
+
+        cmd.CommandText = $"INSERT INTO {tableName} (ingredients) VALUES (@p)";
+        cmd.Parameters.Add(new("p", new[] { jsonObject1, jsonObject2 }));
+        await cmd.ExecuteNonQueryAsync();
     }
 
     public JsonTests(MultiplexingMode multiplexingMode, NpgsqlDbType npgsqlDbType)
